src/requirements.txt
Django==2.2.2
Pillow==6.0.0
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
braintree==3.19.0
defusedxml==0.6.0
django-cors-headers==1.1.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-rest-registration==0.5.6
djangorestframework==3.9.4
djangorestframework-jwt==1.7.2
Markdown==2.6.2
oauthlib==3.1.0
PyJWT==1.7.1
python3-openid==3.1.0
requests==2.7.0
requests-oauthlib==1.3.0
six==1.14.0
typing==3.7.4.1